+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-* pkg: [nodejs](https://nodejs.org); `sync` and `update` install the latest version of Node.js on Linux and macOS
+* pkg: [nodejs](https://nodejs.org); `sync` and `update` install the latest version of Node.js on Linux, macOS, and Windows
 
 ### Fixed